requirements.txt
jupyter==1.0.0
matplotlib==3.8.0
seaborn==0.12.2
numpy==1.26.2
scipy==1.11.4
pandas==2.1.4
soundsig==2.1
scikit-learn==1.2.2
umap-learn==0.5.3
statsmodels==0.11.1
wavio==0.0.4
